% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{assuntos}
\alias{assuntos}
\title{assuntos}
\format{
a data frame with 37022 lines e 16 variables.

\describe{
  \item{tribunal}{indicates which court is concerned.}
  \item{ano}{subject year.}
  \item{assunto_generico}{generic subject.}
  \item{assunto_nome1}{1st branch.}
  \item{assunto_nome2}{2st branch.}
  \item{assunto_nome3}{3st branch.}
  \item{assunto_nome4}{4st branch.}
  \item{assunto_nome5}{5st branch.}
  \item{assunto_nome6}{6st branch.}
  \item{assunto_nome7}{7st branch.}
  \item{codigos}{code related to the subject.}
  \item{assun_cod_inst}{instance codes.}
  \item{x1_grau}{number of cases in 1st degree.}
  \item{x2_grau}{number of cases in 2nd degree.}
  \item{juizado_especial}{number of cases in the special court.}
  \item{turma_especial}{number of cases in the special class.}
  \item{total}{total cases.}
}
}
\source{
\url{https://www.cnj.jus.br/}
}
\usage{
assuntos
}
\description{
A dataset that contains information about hidden ciphers.
}
\examples{

summary(assuntos)
}
\keyword{datasets}
