% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cadmun}
\alias{cadmun}
\title{cadmun}
\format{
um data frame com 5657 e 35 variáveis:
\describe{conjunto de códigos dos municípios e sua situação, contendo
informações sobre fronteira, amazonia, capital, longitude, latitude, entre outros}
}
\source{
cadmun
}
\usage{
cadmun
}
\description{
Um conjunto de dados que contém os códigos de cadastro municipal.
}
\keyword{datasets}
