% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{consumo}
\alias{consumo}
\title{consumo}
\format{
a data frame with 37022 lines e 16 variables.

\describe{
  \item{id_processo}{process identifier number (appeal).}
  \item{assunto}{subject of the process.}
  \item{valor}{value of cause.}
  \item{tipo_litigio}{configuration of the parties to the dispute.}
  \item{dec_val}{second degree decision.}
  \item{dec_unanime}{unanimity.}
  \item{dec_date}{decision date.}
  \item{tempo}{resource time, in days.}
}
}
\source{
\url{https://abj.org.br/}
}
\usage{
consumo
}
\description{
Retrospective basis to use in book examples.
}
\examples{

summary(consumo)
}
\keyword{datasets}
