% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{get_states_map}
\alias{get_states_map}
\title{Gets fortified map from IBGE}
\usage{
get_states_map(path = NULL, type = "state")
}
\arguments{
\item{path}{path of the zip file. If NULL, downloads data from the internet.}

\item{type}{atomic character. Current options are "state" or "city".}
}
\description{
Gets fortified map from IBGE
}
\seealso{
\url{ftp://geoftp.ibge.gov.br/cartas_e_mapas/bases_cartograficas_continuas/bc250/versao2015/Shapefile/}
}

