% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{leiloes}
\alias{leiloes}
\title{Leiloes}
\format{
a data frame with 37022 lines e 16 variables.

\describe{
  \item{id_processo}{process identifier number (appeal).}
  \item{descricao}{item description.}
  \item{id_leiloeiro}{Auctioneer ID.}
  \item{tipo_remuneracao}{Auctioneer compensation type.}
  \item{modalidade}{auction mode.}
  \item{tipo}{auction type.}
  \item{vendeu}{was the item sold?}
  \item{valor_avaliacao_inicial}{appraised value.}
  \item{valor_total_arrematado}{auctioned value.}
}
}
\source{
\url{https://obs.abj.org.br/}
}
\usage{
leiloes
}
\description{
Base on auctions
}
\examples{

summary(leiloes)
}
\keyword{datasets}
