% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{litigiosidade}
\alias{litigiosidade}
\title{Litigation}
\format{
a data frame with 1000 rows and 10 columns.

\describe{
  \item{justica}{Type of justice}
  \item{ano}{Year}
  \item{tribunal_uf}{State acronym}
  \item{baixados}{Closed cases}
  \item{novos}{New cases}
  \item{pendentes}{Pending cases}
  \item{idhm}{State Human Development Index}
  \item{t_analf18m}{Illiteracy rate}
  \item{rdpc}{Per capita income}
  \item{pop}{population}
}
}
\source{
\url{https://abjur.shinyapps.io/obsCIEE/}
}
\usage{
litigiosidade
}
\description{
Litigation data from Brazil.
}
\examples{

summary(litigiosidade)
}
\keyword{datasets}
