% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{muni}
\alias{muni}
\title{muni}
\format{
a data frame with 5572 lines e 16 varibles:

\describe{
  \item{muni_id}{code IBGE (7 digits).}
  \item{muni_id_6}{code (6 digits).}
  \item{muni_nm}{municipality name (original IBGE).}
  \item{muni_nm_clean}{clean municipality name (without accents, upper case).}
  \item{uf_nm}{name of the original UF.}
  \item{uf_sigla}{initials UF.}
  \item{uf_id}{code of IBGE UF.}
  \item{regiao_nm}{region name.}
  \item{tse_id}{municipality TSE code.}
  \item{rf_id}{Receita Federal code of the municipality.}
  \item{bcb_id}{code of the Central Bank of Brazil of the municipality.}
  \item{existia_1991}{1 if municipality existed in 1991, 0 otherwise.}
  \item{existia_2000}{1 if municipality existed in 2000, 0 otherwise.}
  \item{existia_2010}{1 if municipality existed in 2010, 0 otherwise.}
  \item{lon}{longitude based on shapefile centroid.}
  \item{lat}{latitude based on shapefile centroid.}
  \item{capital}{TRUE if UF capital, FALSE otherwise.}
}
}
\source{
geobr package base enriched with TSE and centroids data
}
\usage{
muni
}
\description{
A dataset that contains data useful for joining
with city bases, such as city codes, tse code,
acronym/name/code of the UFs and region.
}
\examples{

summary(muni)
}
\keyword{datasets}
