% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pnud_min}
\alias{pnud_min}
\title{pnud_min}
\format{
um data frame com 16686 linhas e 14 variáveis:
}
\source{
@source \url{https://www.br.undp.org/content/brazil/pt/home/idh0/atlas-do-desenvolvimento-humano/atlas-dos-municipios.html}
}
\usage{
pnud_min
}
\description{
Um conjunto de dados que contém informações sobre PNUD de municípios por anos.
}
\keyword{datasets}
