% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{pnud_muni}
\alias{pnud_muni}
\title{pnud_muni}
\format{
a data frame with 16695 lines e 238 variables.

Considered the year-municipal observational unit.

\describe{more information see the pnud_siglas.}
}
\source{
\url{https://www.br.undp.org/content/brazil/pt/home/idh0/atlas-do-desenvolvimento-humano/atlas-dos-municipios.html}

summary(pnud_muni)
}
\usage{
pnud_muni
}
\description{
A dataset that contains information about UNDP for municipalities and UFs.
}
\keyword{datasets}
