% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{pnud_siglas}
\alias{pnud_siglas}
\title{pnud_siglas}
\format{
a data frame with 237 lines e 4 variables:
\describe{
  \item{sigla}{all acronyms available. }
  \item{nome curto}{short name of the acronym.}
  \item{nome longo}{long name of the acronym.}
  \item{definicao}{definition of acronym.}
}
}
\source{
pnud_siglas
}
\usage{
pnud_siglas
}
\description{
A dataset that serves as a glossary of available acronyms.
}
\examples{

summary(pnud_siglas)
}
\keyword{datasets}
