% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{pnud_uf}
\alias{pnud_uf}
\title{pnud_uf}
\format{
a data frame with 81 lines e 235 variables.

\describe{more information see the pnud_siglas.}
}
\source{
\url{https://www.br.undp.org/content/brazil/pt/home/idh0/rankings/idhm-uf-2010.html}
}
\usage{
pnud_uf
}
\description{
A dataset that contains information about PNUD of Federative Units.
Considered the observational unit year-uf.
}
\examples{

summary(pnud_uf)
}
\keyword{datasets}
