% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{assuntos}
\alias{assuntos}
\title{Case types}
\format{
a data frame with 37022 lines and 15 variables.

\describe{
\item{tribunal}{indicates which court is concerned.}
\item{ano}{subject year.}
\item{assunto_generico}{generic subject.}
\item{assunto_nome1}{1st branch.}
\item{assunto_nome2}{2st branch.}
\item{assunto_nome3}{3st branch.}
\item{assunto_nome4}{4st branch.}
\item{assunto_nome5}{5st branch.}
\item{assunto_nome6}{6st branch.}
\item{assunto_nome7}{7st branch.}
\item{codigos}{code related to the subject.}
\item{assun_cod_inst}{instance codes.}
\item{x1_grau}{number of cases in first instance.}
\item{x2_grau}{number of cases in second instance.}
\item{juizado_especial}{number of cases in the special court.}
\item{turma_especial}{number of cases in the special class.}
\item{total}{total cases.}
}
}
\source{
\url{https://www.cnj.jus.br/}
}
\usage{
assuntos
}
\description{
A dataset contains information about case types in Brazil.
}
\examples{

summary(assuntos)
}
\keyword{datasets}
