% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{assuntos}
\alias{assuntos}
\title{Case types}
\format{
a data frame with 37022 lines and 15 variables.

\describe{
  \item{tribunal}{indicates which court is concerned}
  \item{ano}{subject year}
  \item{assunto_nome1}{first branch}
  \item{assunto_nome2}{second branch}
  \item{assunto_nome3}{third branch}
  \item{assunto_nome4}{fourth branch}
  \item{assunto_nome5}{fifth branch}
  \item{assunto_nome6}{sixth branch}
  \item{generico}{generic code}
  \item{codigo}{code related to the subject}
  \item{x1_grau}{number of cases in first instance}
  \item{x2_grau}{number of cases in second instance}
  \item{juizado_especial}{number of cases in the special court}
  \item{turma_recursal}{number of cases in the special class}
  \item{total}{total cases}
}
}
\source{
\url{https://www.cnj.jus.br/}
}
\usage{
assuntos
}
\description{
A dataset contains information about case types in Brazil.
}
\examples{

summary(assuntos)
}
\keyword{datasets}
