% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{cadmun}
\alias{cadmun}
\title{cadmun}
\format{
a data frame with 5657 lines e 35 variables:

\describe{set of codes of the municipalities and their situation, containing
information about border, amazonia, capital, longitude, latitude, among others}
}
\source{
cadmun
}
\usage{
cadmun
}
\description{
A dataset that contains the municipal cadastre codes.
The database is outdated, as it does not consider all municipalities.
}
\examples{

summary(cadmun)
}
\keyword{datasets}
