% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{cadmun}
\alias{cadmun}
\title{cadmun}
\format{
um data frame com 5657 e 35 variaveis:

\describe{conjunto de codigos dos municipios e sua situacao, contendo
informacoes sobre fronteira, amazonia, capital, longitude, latitude, entre outros}
}
\source{
cadmun
}
\usage{
cadmun
}
\description{
Um conjunto de dados que contem os codigos de cadastro municipal.
A base esta desatualizada, pois nao considera todos os municipios.
}
\keyword{datasets}
