% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{cadmun}
\alias{cadmun}
\title{Municipality info (legacy)}
\format{
a data frame with 5657 rows and 35 variables:

\describe{
  \item{description}{set of codes of the municipalities and their situation, containing information about border, Amazon, capital, longitude, latitude, among others}
}
}
\source{
IBGE
}
\usage{
cadmun
}
\description{
A dataset containing the municipality codes.
The database is outdated, as it does not consider all municipalities.
}
\examples{

summary(cadmun)
}
\keyword{datasets}
