% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{consumo}
\alias{consumo}
\title{Jurisprudencia consumidor}
\format{
um data frame com 37022 linhas e 16 variaveis.

\describe{
  \item{id_processo}{numero identificador do processo (apelacao)}
  \item{assunto}{assunto do processo}
  \item{valor}{valor da causa}
  \item{tipo_litigio}{configuracao das partes no litigio}
  \item{dec_val}{decisao de segundo grau}
  \item{dec_unanime}{unanimidade}
  \item{dec_date}{data da decisao}
  \item{tempo}{tempo do recurso, em dias}
}
}
\usage{
consumo
}
\description{
Base retrospectiva para utilizar em exemplos do livro.
}
\keyword{datasets}
