% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{consumo}
\alias{consumo}
\title{Consumer cases}
\format{
a data frame with 1000 rows and 9 columns.

\describe{
  \item{id_processo}{case identifier number (appeal)}
  \item{assunto}{type of the case}
  \item{comarca}{municipality of the case}
  \item{valor}{value of cause}
  \item{tipo_litigio}{configuration of the parties to the dispute}
  \item{dec_val}{second instance decision}
  \item{dec_unanime}{unanimity}
  \item{dec_date}{decision date}
  \item{tempo}{case time in days}
}
}
\source{
\url{https://abj.org.br/}
}
\usage{
consumo
}
\description{
Retrospective basis to use in book examples.
}
\examples{

summary(consumo)
}
\keyword{datasets}
