% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{leiloes}
\alias{leiloes}
\title{Leiloes}
\format{
um data frame com 37022 linhas e 16 variaveis.

\describe{
  \item{id_processo}{numero identificador do processo (apelacao)}
  \item{descricao}{descricao do item}
  \item{id_leiloeiro}{ID do leiloeiro}
  \item{tipo_remuneracao}{Tipo de remuneração do leiloeiro}
  \item{modalidade}{modalidade do leilão}
  \item{tipo}{tipo de leilão}
  \item{vendeu}{o item foi vendido?}
  \item{valor_avaliacao_inicial}{valor avaliado}
  \item{valor_total_arrematado}{valor arrematado}
}
}
\usage{
leiloes
}
\description{
Base sobre leiloes
}
\keyword{datasets}
