% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{leiloes}
\alias{leiloes}
\title{Auctions}
\format{
a data frame with 1000 rows and 10 columns.

\describe{
  \item{id_processo}{case identifier number}
  \item{descricao}{item description}
  \item{id_leiloeiro}{Auctioneer ID}
  \item{tipo_remuneracao}{Auctioneer compensation type}
  \item{modalidade}{auction mode}
  \item{tipo}{auction type}
  \item{data_edital}{auction date}
  \item{vendeu}{was the item sold?}
  \item{valor_avaliacao_inicial}{appraised value}
  \item{valor_total_arrematado}{auctioned value}
}
}
\source{
\url{https://obs.abj.org.br/}
}
\usage{
leiloes
}
\description{
Auctions sample dataset used in our book.
}
\examples{

summary(leiloes)
}
\keyword{datasets}
