% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{muni}
\alias{muni}
\title{muni}
\format{
um data frame com 5572 linhas e 16 colunas:

\describe{
  \item{muni_id}{codigo IBGE (7 digitos).}
  \item{muni_id_6}{codigo IBGE (6 digitos).}
  \item{muni_nm}{nome do municipio (original IBGE).}
  \item{muni_nm_clean}{nome do municipio limpo (sem acentos, caixa alta).}
  \item{uf_nm}{nome da UF original.}
  \item{uf_sigla}{sigla da UF.}
  \item{uf_id}{codigo IBGE da UF.}
  \item{regiao_nm}{nome da regiao.}
  \item{tse_id}{codigo TSE do municipio.}
  \item{rf_id}{codigo Receita Federal do municipio.}
  \item{bcb_id}{codigo do Banco Central do Brasil do municipio.}
  \item{existia_1991}{1 se municipio existia em 1991, 0 caso contrario.}
  \item{existia_2000}{1 se municipio existia em 2000, 0 caso contrario.}
  \item{existia_2010}{1 se municipio existia em 2010, 0 caso contrario.}
  \item{lon}{longitude baseada no centroide do shapefile.}
  \item{lat}{latitude baseada no centroide do shapefile.}
}
}
\source{
base do pacote geobr enriquecida com dados do TSE e centroides
}
\usage{
muni
}
\description{
Um conjunto de dados que contem dados uteis para fazer join
com bases de municipios, como codigos dos municipios, codigo tse,
sigla/nome/codigo das UFs e regiao.
}
\keyword{datasets}
