% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{pnud_min}
\alias{pnud_min}
\title{pnud_min}
\format{
a data frame woth 16686 lines e 14 variables.

\describe{more information see the pnud_siglas.}
}
\source{
\url{https://www.br.undp.org/content/brazil/pt/home/idh0/atlas-do-desenvolvimento-humano/atlas-dos-municipios.html}
}
\usage{
pnud_min
}
\description{
A dataset that contains PNUD information for municipalities by years.
}
\examples{

summary(pnud_siglas)
}
\keyword{datasets}
