% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{pnud_min}
\alias{pnud_min}
\title{UNDP minimal dataset}
\format{
a data frame with 16686 rows and 15 columns.

\describe{
  \item{ano}{for more information, check \code{\link{pnud_siglas}}}
}
}
\source{
\url{https://www.br.undp.org/content/brazil/pt/home/idh0/atlas-do-desenvolvimento-humano/atlas-dos-municipios.html}
}
\usage{
pnud_min
}
\description{
A dataset that contains UNDP information for municipalities by years.
}
\examples{

summary(pnud_min)
}
\keyword{datasets}
