% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pnud_muni}
\alias{pnud_muni}
\title{pnud_muni}
\format{
um data frame com 16695 linhas e 238 variáveis:
\describe{mais informações consulte a pnud_siglas}
}
\source{
\url{https://www.br.undp.org/content/brazil/pt/home/idh0/atlas-do-desenvolvimento-humano/atlas-dos-municipios.html}
}
\usage{
pnud_muni
}
\description{
Um conjunto de dados que contém informações sobre PNUD de municípios.
}
\keyword{datasets}
