% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{pnud_siglas}
\alias{pnud_siglas}
\title{pnud_siglas}
\format{
um data frame com 237 e 4 variaveis:
\describe{
  \item{sigla}{todas as siglas disponiveis}
  \item{nome curto}{nome curto da sigla}
  \item{nome longo}{nome longo da sigla}
  \item{definicao}{definicao da sigla}
}
}
\source{
pnud_siglas
}
\usage{
pnud_siglas
}
\description{
Um conjuto de dados que serve como glossario das siglas disponiveis.
}
\keyword{datasets}
