% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{pnud_siglas}
\alias{pnud_siglas}
\title{UNDP Acronyms}
\format{
a data frame with 236 rows e 4 columns:

\describe{
  \item{sigla}{all acronyms available}
  \item{nome_curto}{short name of the variable}
  \item{nome_longo}{long name of the variable}
  \item{definicao}{definition of acronym}
}
}
\source{
\url{https://www.br.undp.org/content/brazil/pt/home/idh0/atlas-do-desenvolvimento-humano/atlas-dos-municipios.html}
}
\usage{
pnud_siglas
}
\description{
A dataset that serves as a glossary of available variables.
}
\examples{

summary(pnud_siglas)
}
\keyword{datasets}
