% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pnud_uf}
\alias{pnud_uf}
\title{pnud_uf}
\format{
um data frame com 81 linhas e 235 variáveis:
\describe{mais informações consulte a pnud_siglas}
}
\source{
\url{https://www.br.undp.org/content/brazil/pt/home/idh0/rankings/idhm-uf-2010.html}
}
\usage{
pnud_uf
}
\description{
Um conjunto de dados que contém informações sobre PNUD de Unidades Federativas.
}
\keyword{datasets}
