% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{pnud_uf}
\alias{pnud_uf}
\title{UNDP data by Federative Units}
\format{
a data frame with 81 rows e 235 columns.

\describe{
  \item{ano}{for more information, check \code{\link{pnud_siglas}}}
}
}
\source{
\url{https://www.br.undp.org/content/brazil/pt/home/idh0/rankings/idhm-uf-2010.html}
}
\usage{
pnud_uf
}
\description{
A dataset that contains information about UNDP of Federative Units.
}
\examples{

summary(pnud_uf)
}
\keyword{datasets}
